Revert "Update flask structure"
