--- api/requirements.txt
@@ -1,2 +1,0 @@
-Flask==3.1.0
-python-dotenv==1.0.1
